Correção no End-Point no finalizar venda

+++ APIDevSteamJau/Controllers/ItensCarrinhosController.cs
@@ -116,6 +116,23 @@ namespace APIDevSteamJau.Controllers
                 return NotFound();
             }
 
+            //Verificar se o carrinho existe
+            var carrinho = await _context.Carrinhos.FindAsync
+                (itemCarrinho.CarrinhoId);
+            if (carrinho == null)
+            {
+                return NotFound("Carrinho nao encontrado.");
+            }
+
+            //Remove o valor total do carrinho
+            carrinho.ValorTotal -= itemCarrinho.ValorTotal;
+
+            //Verifica se o valor do carrinho é menor que zero
+            if (carrinho.ValorTotal < 0)
+            {
+                carrinho.ValorTotal = 0;
+            }
+
             _context.ItensCarrinhos.Remove(itemCarrinho);
             await _context.SaveChangesAsync();
